+++ app/Exceptions/Demand/AdPayReportNotReadyException.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (c) 2018 Adshares sp. z o.o.
+ * Copyright (c) 2018-2019 Adshares sp. z o.o.
  *
  * This file is part of AdServer
  *
@@ -18,29 +18,10 @@
  * along with AdServer. If not, see <https://www.gnu.org/licenses/>
  */
 
-declare(strict_types = 1);
+namespace Adshares\Adserver\Exceptions\Demand;
 
-namespace Adshares\Demand\Application\Service;
+use Adshares\Common\Exception\RuntimeException;
 
-use Adshares\Demand\Application\Dto\AdPayEvents;
-
-interface AdPay
+class AdPayReportNotReadyException extends RuntimeException
 {
-    public function updateCampaign(array $campaigns): void;
-
-    public function deleteCampaign(array $campaignIds): void;
-
-    public function addViews(AdPayEvents $events): void;
-
-    public function addClicks(AdPayEvents $events): void;
-
-    public function addConversions(AdPayEvents $events): void;
-
-    public function getPayments(
-        int $timestamp,
-        bool $recalculate = false,
-        bool $force = false,
-        ?int $limit = null,
-        ?int $offset = null
-    ): array;
 }