+++ app/Models/ConversionDefinition.php
@@ -113,6 +113,11 @@ class ConversionDefinition extends Model
         'creating' => GenerateUUID::class,
     ];
 
+    public static function fetchById(int $id): ?self
+    {
+        return self::find($id);
+    }
+
     public static function fetchByUuid(string $uuid): ?self
     {
         return self::where('uuid', hex2bin($uuid))->first();
@@ -125,7 +130,7 @@ class ConversionDefinition extends Model
 
     public function conversionGroups(): HasMany
     {
-        return $this->hasMany(ConversionGroup::class);
+        return $this->hasMany(Conversion::class);
     }
 
     public function getLinkAttribute()
@@ -154,9 +159,23 @@ class ConversionDefinition extends Model
         return self::ADVANCED_TYPE === $this->type;
     }
 
-    public function isRepeatable(): bool
+    public function isInCampaignBudget(): bool
     {
-        return (bool)$this->is_repeatable;
+        return self::IN_BUDGET === $this->limit_type;
+    }
+
+    public static function updateCostAndOccurrences(array $costAndOccurrencesArray): void
+    {
+        $ids = array_keys($costAndOccurrencesArray);
+        $definitions = ConversionDefinition::whereIn('id', $ids)->get();
+
+        foreach ($definitions as $definition) {
+            $data = $costAndOccurrencesArray[$definition->id];
+
+            $definition->cost = $data['cost'];
+            $definition->occurrences = $data['occurrences'];
+            $definition->save();
+        }
     }
 
     public static function removeFromCampaignWithoutGivenUuids(int $campaignId, array $uuids): void
@@ -175,7 +194,7 @@ class ConversionDefinition extends Model
     public static function rules(array $conversion): array
     {
         $type = $conversion['type'] ?? null;
-        $isValueMutable = $conversion['is_value_mutable'] ?? null;
+        $isValueMutable = (bool)($conversion['is_value_mutable'] ?? false);
         $rules = [
             'uuid' => 'string|nullable',
             'campaign_id' => 'required|integer',
@@ -187,7 +206,7 @@ class ConversionDefinition extends Model
                 'min:0',
                 'nullable',
                 Rule::requiredIf(static function () use ($isValueMutable) {
-                    return 0 === $isValueMutable;
+                    return !$isValueMutable;
                 }),
             ],
             'limit' => 'integer|min:0|nullable',
@@ -195,12 +214,18 @@ class ConversionDefinition extends Model
 
         if ($type === self::BASIC_TYPE) {
             $rules['limit_type'] = 'required|in:'.self::IN_BUDGET;
-            $rules['is_repeatable'] = 'required|in:0';
-            $rules['is_value_mutable'] = 'required|in:0';
+            $rules['is_repeatable'] = [
+                'required',
+                Rule::in(false, 0),
+            ];
+            $rules['is_value_mutable'] = [
+                'required',
+                Rule::in(false, 0),
+            ];
         } elseif ($type === self::ADVANCED_TYPE) {
             $rules['limit_type'] = sprintf('required|in:%s,%s', self::IN_BUDGET, self::OUT_OF_BUDGET);
-            $rules['is_repeatable'] = 'required|in:0,1';
-            $rules['is_value_mutable'] = 'required|in:0,1';
+            $rules['is_repeatable'] = 'required|boolean';
+            $rules['is_value_mutable'] = 'required|boolean';
         }
 
         return $rules;