+++ app/Providers/Common/ClientProvider.php
@@ -24,6 +24,7 @@ namespace Adshares\Adserver\Providers\Common;
 
 use Adshares\Ads\AdsClient;
 use Adshares\Adserver\Client\ClassifierExternalClient;
+use Adshares\Adserver\Client\GuzzleAdPayClient;
 use Adshares\Adserver\Client\GuzzleAdSelectLegacyClient;
 use Adshares\Adserver\Client\GuzzleAdSelectClient;
 use Adshares\Adserver\Client\GuzzleAdsOperatorClient;
@@ -31,7 +32,7 @@ use Adshares\Adserver\Client\GuzzleAdUserClient;
 use Adshares\Adserver\Client\GuzzleClassifierExternalClient;
 use Adshares\Adserver\Client\GuzzleDemandClient;
 use Adshares\Adserver\Client\GuzzleLicenseClient;
-use Adshares\Adserver\Client\JsonRpcAdPayClient;
+use Adshares\Adserver\Client\JsonRpcAdPayLegacyClient;
 use Adshares\Adserver\Client\JsonRpcAdSelectLegacyClient;
 use Adshares\Adserver\Client\LocalPublisherBannerClassifier;
 use Adshares\Adserver\Client\MultipleExternalClassifierAdClassifyClient;
@@ -48,6 +49,7 @@ use Adshares\Common\Application\Service\LicenseProvider;
 use Adshares\Common\Application\Service\SignatureVerifier;
 use Adshares\Common\Infrastructure\Service\PhpAdsClient;
 use Adshares\Demand\Application\Service\AdPay;
+use Adshares\Demand\Application\Service\AdPayLegacy;
 use Adshares\Supply\Application\Service\AdSelectLegacy;
 use Adshares\Supply\Application\Service\AdSelect;
 use Adshares\Supply\Application\Service\BannerClassifier;
@@ -62,9 +64,9 @@ final class ClientProvider extends ServiceProvider
     public function register(): void
     {
         $this->app->bind(
-            AdPay::class,
+            AdPayLegacy::class,
             function () {
-                return new JsonRpcAdPayClient(
+                return new JsonRpcAdPayLegacyClient(
                     new JsonRpc(
                         new Client(
                             [
@@ -73,6 +75,21 @@ final class ClientProvider extends ServiceProvider
                                 'timeout' => 300,
                             ]
                         )
+                    )
+                );
+            }
+        );
+
+        $this->app->bind(
+            AdPay::class,
+            function () {
+                return new GuzzleAdPayClient(
+                    new Client(
+                        [
+                            'headers' => ['Content-Type' => 'application/json', 'Cache-Control' => 'no-cache'],
+                            'base_uri' => config('app.adpay_endpoint'),
+                            'timeout' => 300,
+                        ]
                     )
                 );
             }