+++ CHANGELOG.md
@@ -7,13 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- NetworkEventLog and related classes
 - Banner checksum fragment to serve url
+- Exporting conversions to adpay
 ### Changed
 - Sum incomes/expenditures from different adservers in billing history
+- Getting payment report from adpay (optimization)
 ### Fixed
 - Grouping events by day in billing history
 - Importing banners when external classifier is not defined
+### Removed
+- NetworkEventLog and related classes
 
 ## [1.2.5] - 2019-10-01
 ### Added